Remove some RESET calls and add undefined checks. Fixes #142.
Removed redundant RESET calls in SAVE_COHORT; the call to SET_COHORT takes care of several
of them. The call to RESET_CROSS_FILTER was removed completely because it resets the cross
filter back to null, which is overkill and results in exceptions down the line. Also
added a check in actions.js to ensure that filterAll() isn't called on an undefined
dimension. Not clear why this is happening but the combination of these workarounds
appears to fix the problem.